feat: add support for YAML when validating and inspecting
